Moved the compact_partition tests to top level module
